requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
numpy
xarray
rioxarray
rasterio
pystac-client
planetary-computer
odc-stac
torch
transformers
safetensors
scikit-image
httpx
